+++ QSymbolicsBase/CHANGELOG.md
@@ -1,5 +1,7 @@
 # News
 
-## v0.1.1 - 2023-01-31
+## v0.1.1 - dev
 
 - Documentation and more symbol names available for the qubit base states.
+- `SApplyOp` implemented for superoperator acting on operator.
+- `STensorSuperOperator` implemented for tensor products of superoperators.