+++ QSymbolicsOptics/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## v0.1.1 - dev
 
-- Documentation and internal cleanup of the qubit bases
+- Documentation and internal cleanup of the qubit bases.
+- `LazySuperTensor` implemented (should be moved to `QuantumOpticsBase`).